reimplement advanced tickers
