--- src/web/ClientApp/src/app/stocks/stock-trading/stock-position-reports.component.ts
@@ -1,35 +1,0 @@
-import { Component, Input } from '@angular/core';
-import { StocksService, StockGaps } from '../../services/stocks.service';
-
-@Component({
-  selector: 'app-position-reports',
-  templateUrl: './stock-position-reports.component.html',
-  styleUrls: ['./stock-position-reports.component.css']
-})
-export class StockPositionReportsComponent {
-
-  
-  sortColumn: string
-  sortDirection: number = -1
-  
-
-	constructor(private service : StocksService){}
-
-  @Input()
-  dailyMode: boolean = false
-
-  @Input()
-  allTimeMode: boolean = false
-
-
-	ngOnInit(): void {
-    
-    if (this.allTimeMode) {
-      // this.loadAllTimeData()
-    }
-
-    if (this.dailyMode) {
-      // this.loadDailyData()
-    }
-  }
-}
