+++ src/web/ClientApp/src/app/stocks/stock-trading/stock-trading-dashboard.component.html
@@ -77,7 +77,7 @@
   <app-stock-position-charts  *ngIf="isActive('openreview')" [positions]=positions></app-stock-position-charts>
   <app-stock-violations       *ngIf="isActive('violations')" [violations]="violations" (refreshRequested)="refresh()"></app-stock-violations>
   
-  <app-position-reports       *ngIf="isActive('analysis')" [allTimeMode]="true"></app-position-reports>
-  <app-position-reports       *ngIf="isActive('dailyanalysis')" [dailyMode]="true"></app-position-reports>
+  <app-position-reports       *ngIf="isActive('analysis')" [allTimeMode]="true" [positions]="positions"></app-position-reports>
+  <app-position-reports       *ngIf="isActive('dailyanalysis')" [dailyMode]="true" [positions]="positions"></app-position-reports>
 
 </div>
