+++ src/web/ClientApp/src/app/stocks/stock-trading/stock-trading-outcomes-reports.component.css
