+++ src/web/ClientApp/src/app/stocks/stock-trading/stock-trading-outcomes-reports.component.html
@@ -1,5 +1,10 @@
 <div>
-  <app-outcomes *ngIf="multipleBarOutcomes" [title]="'Mutliple bar outcomes'" [outcomes]="multipleBarOutcomes"></app-outcomes>
+
+  <app-outcomes
+    *ngIf="allBarsReport"
+    [title]="'Mutliple bar outcomes'"
+    [outcomes]="allBarsReport.outcomes">
+  </app-outcomes>
 
   <app-outcomes-analysis-report
     *ngIf="positionReport"
@@ -28,6 +33,6 @@
     [outcomes]="dailyOutcomes">
   </app-outcomes>
 
-  <app-gaps *ngFor="let gap of gaps" [gaps]="gap"></app-gaps>
+  <app-gaps *ngFor="let gap of allBarsReport.gaps" [gaps]="gap"></app-gaps>
   
 </div>
